+++ levelup-react-16/src/Movie.js
@@ -1,11 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types'
+import {Link} from 'react-router-dom'
 
 const POSTER_PATH = 'http://image.tmdb.org/t/p/w154';
 
  const Movie = ({movie}) => (
              <div>
+                 <Link to={`/${movie.id}`}>
                 <img src ={`${POSTER_PATH}${movie.poster_path}`} alt = {movie.title} />
+                </Link>
             </div>
         
  )
